# trace: Improve support for batching field values