--- a/tokio-trace/src/span.rs
+++ b/tokio-trace/src/span.rs
@@ -146,7 +146,6 @@ use std::{
 use {
     dispatcher::{self, Dispatch},
     field,
-    subscriber::Interest,
     Metadata,
 };