--- a/tokio-trace/tests/support/mod.rs
+++ b/tokio-trace/tests/support/mod.rs
@@ -1,3 +1,4 @@
 #![allow(dead_code)]
+pub mod field;
 pub mod span;
 pub mod subscriber;